refactor scroll logic
